requirements.txt
# OpenAI Whisper - Transcription locale
openai-whisper

# Audio
pyaudio